# specs/001-saas-mvp/contracts/admin-triage.openapi.yaml
openapi: 3.1.0
info:
  title: ARGUS Admin & Triage REST API
  version: 1.0.0
  description: |
    Admin Dashboard configuration API and Triage SPA operational API.
    Separate from edge ingestion API (see edge-ingestion.openapi.yaml).

servers:
  - url: https://api.argus.example.com/v1

paths:
  # ── Root Admin ──────────────────────────────────────────
  /admin/tenants:
    post:
      operationId: createTenant
      summary: Create a new tenant (Root Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
    get:
      operationId: listTenants
      summary: List all tenants (Root Admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tenant list

  /admin/tenants/{tenant_id}/admins:
    post:
      operationId: assignTenantAdmin
      summary: Assign Tenant Admin to tenant
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, idp_subject]
              properties:
                email:
                  type: string
                idp_subject:
                  type: string
      responses:
        '201':
          description: Admin assigned

  # ── Tenant Admin: Markets & Cameras ─────────────────────
  /tenants/{tenant_id}/markets:
    get:
      operationId: listMarkets
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Market list
    post:
      operationId: createMarket
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMarketRequest'
      responses:
        '201':
          description: Market created

  /tenants/{tenant_id}/markets/{market_id}/cameras:
    post:
      operationId: createCamera
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/MarketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCameraRequest'
      responses:
        '201':
          description: Camera registered

  /tenants/{tenant_id}/cameras/{camera_id}/regions:
    post:
      operationId: createRegion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/CameraId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegionRequest'
      responses:
        '201':
          description: Region created

  # ── Context Modes & Rules ───────────────────────────────
  /tenants/{tenant_id}/context-modes:
    post:
      operationId: createContextMode
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContextModeRequest'
      responses:
        '201':
          description: Context mode created

  /tenants/{tenant_id}/context-modes/{mode_id}/schedules:
    post:
      operationId: createSchedule
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: mode_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: Schedule created

  /tenants/{tenant_id}/context-modes/{mode_id}/lenses:
    post:
      operationId: createLens
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: mode_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLensRequest'
      responses:
        '201':
          description: Lens created

  /tenants/{tenant_id}/rules:
    post:
      operationId: createRule
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
      responses:
        '201':
          description: Rule created

  # ── Notifications ─────────────────────────────────────
  /tenants/{tenant_id}/notification-configs:
    post:
      operationId: createNotificationConfig
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationConfigRequest'
      responses:
        '201':
          description: Config created

  # ── Triage (Watcher) ──────────────────────────────────
  /tenants/{tenant_id}/decisions:
    get:
      operationId: listDecisions
      summary: List active Decisions for triage board
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: state
          in: query
          schema:
            type: string
            enum: [normal, weird, warning, resolved]
      responses:
        '200':
          description: Decision list

  /tenants/{tenant_id}/decisions/{decision_id}:
    get:
      operationId: getDecision
      summary: Decision detail with linked evidences
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/DecisionId'
      responses:
        '200':
          description: Decision detail

  /tenants/{tenant_id}/decisions/{decision_id}/resolve:
    post:
      operationId: resolveDecision
      summary: Watcher resolves a Warning Decision
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/DecisionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveDecisionRequest'
      responses:
        '200':
          description: Decision resolved
        '409':
          description: Already resolved by another Watcher

  /tenants/{tenant_id}/decisions/{decision_id}/evidences/{evidence_id}/playback:
    get:
      operationId: getEvidencePlayback
      summary: Presigned URL for frame sequence playback
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/DecisionId'
        - name: evidence_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Presigned playback URL

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    TenantId:
      name: tenant_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MarketId:
      name: market_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CameraId:
      name: camera_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DecisionId:
      name: decision_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    CreateTenantRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        aggregation_window_secs:
          type: integer
          default: 300

    CreateMarketRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        timezone:
          type: string
          default: UTC

    CreateCameraRequest:
      type: object
      required: [name, edge_device_id]
      properties:
        name:
          type: string
        edge_device_id:
          type: string

    CreateRegionRequest:
      type: object
      required: [name, polygon]
      properties:
        name:
          type: string
        polygon:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number

    CreateContextModeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string

    CreateScheduleRequest:
      type: object
      required: [day_of_week, start_time, end_time]
      properties:
        day_of_week:
          type: string
          enum: [mon, tue, wed, thu, fri, sat, sun]
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        market_id:
          type: string
          format: uuid

    CreateLensRequest:
      type: object
      required: [name, system_prompt, output_schema]
      properties:
        name:
          type: string
        system_prompt:
          type: string
        output_schema:
          type: object

    CreateRuleRequest:
      type: object
      required: [context_mode_id, name, condition, region_ids]
      properties:
        context_mode_id:
          type: string
          format: uuid
        name:
          type: string
        condition:
          type: object
        severity_weight:
          type: integer
          default: 1
        region_ids:
          type: array
          items:
            type: string
            format: uuid

    CreateNotificationConfigRequest:
      type: object
      required: [channel, recipient]
      properties:
        channel:
          type: string
          enum: [sms, whatsapp]
        recipient:
          type: string
          description: E.164 phone number

    ResolveDecisionRequest:
      type: object
      required: [disposition]
      properties:
        disposition:
          type: string
          enum: [true_positive, false_positive, false_negative]
        reasoning:
          type: string
          description: Required when disposition is false_positive or false_negative